Update and finish level map render system
- Implement rendering of specific layers of a level
- Render only visible part of the level map
